Rename endpoint files to better represent the actual ESI endpoint names
