+++ src/endpoints/mod.rs
@@ -1,28 +1,28 @@
 use crate::EsiClient;
 
-pub mod alliance;
-pub mod character;
-pub mod corporation;
+pub mod alliances;
+pub mod characters;
+pub mod corporations;
 
 impl EsiClient {
     /// Access to Alliance ESI endpoints
     ///
     /// Returns an API client for interacting with alliance-related endpoints.
-    pub fn alliances(&self) -> self::alliance::AllianceApi {
-        self::alliance::AllianceApi::new(self)
+    pub fn alliances(&self) -> self::alliances::AllianceApi {
+        self::alliances::AllianceApi::new(self)
     }
 
     /// Access to Character ESI endpoints
     ///
     /// Returns an API client for interacting with character-related endpoints.
-    pub fn characters(&self) -> self::character::CharacterApi {
-        self::character::CharacterApi::new(self)
+    pub fn characters(&self) -> self::characters::CharacterApi {
+        self::characters::CharacterApi::new(self)
     }
 
     /// Access to Corporation ESI endpoints
     ///
     /// Returns an API client for interacting with corporation-related endpoints.
-    pub fn corporations(&self) -> self::corporation::CorporationApi {
-        self::corporation::CorporationApi::new(self)
+    pub fn corporations(&self) -> self::corporations::CorporationApi {
+        self::corporations::CorporationApi::new(self)
     }
 }
